backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
meteostat==1.6.5
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
apscheduler==3.10.4
python-multipart==0.0.6
httpx==0.25.2 